refactor : #89 함수 재사용을 위한 리팩토링
DateString 함수 분리 및 재사용
