+++ project_frontend/src/components/boardDetails/boardContent.js
@@ -16,21 +16,13 @@ import {
 	ReplySubmitButton,
 	ButtonsWrap,
 	BoardContentButtonDiv,
-} from './boardDetailsStyledComponents';
+} from './StyledComponents';
 import {
 	getRecruitment,
 	closeRecruitment,
 	deleteRecruitment,
 } from '../../apis';
-
-const DateString = (dateString, parseString) => {
-	let date = new Date(dateString);
-	var year = date.getFullYear();
-	var month = ('0' + (date.getMonth() + 1)).slice(-2);
-	var day = ('0' + date.getDate()).slice(-2);
-	var newDateString = year + parseString + month + parseString + day;
-	return newDateString;
-};
+import { DateString } from '../../utils';
 
 const BoardContent = () => {
 	const params = useParams();
