+++ project_frontend/src/utils/index.js
@@ -1,4 +1,5 @@
 import { movePath } from './movePath';
 import { isLogined } from './isLogined';
+import { DateString } from './dateString';
 
-export { movePath, isLogined };
+export { movePath, isLogined, DateString };
